refactor: update file and test names
