+++ cmd/osv-scanner/scan/image/command.go
