+++ cmd/osv-scanner/update/command.go
